Move all mapping to CoreDataMapper
